role & seeder

+++ app/Http/Controllers/ProductController.php
@@ -4,40 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-// use DataTables;
 
 class ProductController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index(Request $request)
     {
         $products = Product::latest()->paginate(10);
         return view('product', compact('products'));
     }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
+        $this->authorize('manage products');
         $request->validate(
             [
                 'name'=>'required',
@@ -62,39 +39,9 @@ class ProductController extends Controller
             'status'=>'success',
         ]);
     }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        $product = Product::find($id);
-        return response()->json($product);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request)
     {
+        $this->authorize('manage products');
         $request->validate(
             [
                 'up_name'=>'required',
@@ -119,28 +66,20 @@ class ProductController extends Controller
             'status'=>'success',
         ]);
     }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Request $request)
     {
+        $this->authorize('manage products');
         Product::find($request->product_id)->delete();
       
         return response()->json([
             'status'=>'success'
         ]);
     }
-
     public function pagination(Request $request)
     {
         $products = Product::latest()->paginate(10);
         return view('pagination_products', compact('products'))->render();
     }
-
     public function searchProduct(Request $request)
     {
         $products = Product::where('name', 'like', '%'.$request->search_string.'%')
